Split timing code out into its own file
